refactor(frontend): Restructure app component
