+++ frontend/src/lib/components/apps/editor/GridEditor.svelte
@@ -9,7 +9,7 @@
 	import RecomputeAllComponents from './RecomputeAllComponents.svelte'
 	import type { Policy } from '$lib/gen'
 	import HiddenComponent from '../components/helpers/HiddenComponent.svelte'
-	import Component from './Component.svelte'
+	import Component from './component/Component.svelte'
 
 	export let policy: Policy
