+++ frontend/src/lib/components/apps/editor/component/index.ts
@@ -1,0 +1,5 @@
+export { default as Component } from './Component.svelte'
+
+export * from './components'
+export * from './default-codes'
+export * from './sets'
